chain/ethereum,graph,node: moved BlockIngestor to chain/ethereum and removed IngestorAdapter
The IngestorAdapter was actually protocol specific and having the abstraction in the Blockchain trait was not used expect for Ethereum chain.

Removed `IngestorAdapter` from block chain and moved the chain agnostic `BlockIngestor` to `graph-chain-ethereum` crate.
