+++ graph/src/blockchain/mod.rs
@@ -2,7 +2,6 @@
 //! blockchain into Graph Node. A blockchain is represented by an implementation of the `Blockchain`
 //! trait which is the centerpiece of this module.
 
-pub mod block_ingestor;
 pub mod block_stream;
 pub mod firehose_block_ingestor;
 pub mod firehose_block_stream;
@@ -98,8 +97,6 @@ pub trait Blockchain: Debug + Sized + Send + Sync + Unpin + 'static {
 
     type NodeCapabilities: NodeCapabilities<Self> + std::fmt::Display;
 
-    type IngestorAdapter: IngestorAdapter<Self>;
-
     type RuntimeAdapter: RuntimeAdapter<Self>;
 
     fn triggers_adapter(
@@ -129,8 +126,6 @@ pub trait Blockchain: Debug + Sized + Send + Sync + Unpin + 'static {
         metrics: Arc<BlockStreamMetrics>,
         unified_api_version: UnifiedMappingApiVersion,
     ) -> Result<Box<dyn BlockStream<Self>>, Error>;
-
-    fn ingestor_adapter(&self) -> Arc<Self::IngestorAdapter>;
 
     fn chain_store(&self) -> Arc<dyn ChainStore>;
 
@@ -171,39 +166,6 @@ impl From<Error> for IngestorError {
 impl From<web3::Error> for IngestorError {
     fn from(e: web3::Error) -> Self {
         IngestorError::Unknown(anyhow::anyhow!(e))
-    }
-}
-
-#[async_trait]
-pub trait IngestorAdapter<C: Blockchain> {
-    fn logger(&self) -> &Logger;
-
-    /// How many ancestors of the current chain head to ingest. For chains
-    /// that can experience reorgs, this should be large enough to cover all
-    /// blocks that could be subject to reorgs to ensure that `graph-node`
-    /// has enough blocks in its local cache to traverse a sidechain back to
-    /// the main chain even if those blocks get removed from the network
-    /// client.
-    fn ancestor_count(&self) -> BlockNumber;
-
-    /// Get the latest block from the chain
-    async fn latest_block(&self) -> Result<BlockPtr, IngestorError>;
-
-    /// Retrieve all necessary data for the block  `hash` from the chain and
-    /// store it in the database
-    async fn ingest_block(&self, hash: &BlockHash) -> Result<Option<BlockHash>, IngestorError>;
-
-    /// Return the chain head that is stored locally, and therefore visible
-    /// to the block streams of subgraphs
-    fn chain_head_ptr(&self) -> Result<Option<BlockPtr>, Error>;
-
-    /// Remove old blocks from the database cache and return a pair
-    /// containing the number of the oldest block retained and the number of
-    /// blocks deleted if anything was removed. This is generally only used
-    /// in small test installations, and can remain a noop without
-    /// influencing correctness.
-    fn cleanup_cached_blocks(&self) -> Result<Option<(i32, usize)>, Error> {
-        Ok(None)
     }
 }
